feat: 이미지 저장 기능 구현
