+++ apps/client/package.json
@@ -15,6 +15,8 @@
     "@vanilla-extract/css": "^1.17.0",
     "@vanilla-extract/recipes": "^0.5.5",
     "axios": "^1.7.9",
+    "dom-to-image": "^2.6.0",
+    "file-saver": "^2.0.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router-dom": "^7.1.1"
@@ -23,6 +25,8 @@
     "@confeti/eslint": "workspace:*",
     "@confeti/typescript": "workspace:*",
     "@types/axios": "^0.14.4",
+    "@types/dom-to-image": "^2.6.7",
+    "@types/file-saver": "^2.0.7",
     "@types/react": "19.0.0",
     "@types/react-dom": "19.0.1",
     "@vanilla-extract/vite-plugin": "^4.0.19",
