+++ apps/client/src/pages/time-table/components/time-table-board/time-table-board.css.ts
@@ -27,3 +27,11 @@ export const timeBar = style({
   width: '100%',
   color: themeVars.color.gray300,
 });
+
+export const bottomWrapper = style({
+  padding: '2rem 2rem 0rem',
+});
+
+export const saveButton = style({
+  borderRadius: '5px',
+});
