+++ apps/client/src/pages/time-table/components/time-table-section/time-table-section.css.ts
@@ -3,11 +3,3 @@ import { style } from '@vanilla-extract/css';
 export const container = style({
   padding: '4.2rem 0rem 2rem',
 });
-
-export const bottomWrapper = style({
-  padding: '2rem 2rem 0rem',
-});
-
-export const saveButton = style({
-  borderRadius: '5px',
-});
